Added some examples.
Each animation corresponds to a different temperatura, as specified in the title. For T=0 one expects entropy to be low, and the system did in fact get to a metastable state. When the temperature goes up, it can be seen that the entropy rapidly increases —more so for higher temperatures.